requirements.txt
garminconnect==0.2.8
gspread==6.2.1
python-dotenv==1.2.1
